pyproject.toml
[tool.poetry]
name = "pysonolus"
version = "0.1.0"
description = "Sonolus Developer Toolkit in Python"
authors = ["忘忧北萱草 <wybxc@qq.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
sourceinspect = "^0.0.4"
sonolus-core-py = {path = "../sonolus-core-py"}

[tool.poetry.dev-dependencies]
yapf = "^0.32.0"
toml = "^0.10.2"
isort = "^5.10.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
